refactored components and created favorite bar component
