+++ src/components/FavoriteBar/styles.js
