--- src/stories/components/FavoriteBar/FavoriteBar.stories.js
@@ -1,4 +1,0 @@
-import React from 'react'
-
-import FavoriteBar from '../../../components/FavoriteBar/FavoriteBar'
-
